+++ src/vorbis.c
@@ -4,7 +4,7 @@
  */
 #include "tagspriv.h"
 
-#define leuint(d) (((uchar*)(d))[3]<<24 | ((uchar*)(d))[2]<<16 | ((uchar*)(d))[1]<<8 | ((uchar*)(d))[0]<<0)
+#define leuint(d) (uint)(((uchar*)(d))[3]<<24 | ((uchar*)(d))[2]<<16 | ((uchar*)(d))[1]<<8 | ((uchar*)(d))[0]<<0)
 
 void
 cbvorbiscomment(Tagctx *ctx, char *k, char *v){
